# openapi.yaml
openapi: 3.0.3
info:
  title: Reconciliation API
  version: RECONCILIATION_VERSION
paths:
...
  /rules:
    post:
      summary: Create a rule
      operationId: createRule
      tags: [reconciliation.v1]
      requestBody: { $ref: '#/components/requestBodies/Rule' }
      responses:
        '201': { $ref: '#/components/responses/Rule' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
    get:
      summary: List rules
      operationId: listRules
      tags: [reconciliation.v1]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueryBuilder' }
      responses:
        '200': { $ref: '#/components/responses/Rules' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:read]
  /rules/{ruleID}:
    parameters: [{ $ref: '#/components/parameters/RuleID' }]
    get:
      summary: Get a rule
      operationId: getRule
      tags: [reconciliation.v1]
      responses:
        '200': { $ref: '#/components/responses/Rule' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:read]
    patch:
      summary: Patch a rule (partial update)
      operationId: patchRule
      tags: [reconciliation.v1]
      requestBody: { $ref: '#/components/requestBodies/RulePatch' }
      responses:
        '200': { $ref: '#/components/responses/Rule' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
    delete:
      summary: Delete a rule definition while preserving its audit history
      operationId: deleteRule
      tags: [reconciliation.v1]
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
  /rules/{ruleID}/evaluate:
    post:
      summary: Evaluate a rule now
      operationId: evaluateRule
      tags: [reconciliation.v1]
      parameters: [{ $ref: '#/components/parameters/RuleID' }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EvaluateRuleRequest' }
      responses:
        '200': { $ref: '#/components/responses/Evaluation' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
...
components:
  parameters:
    PageSize:
      name: pageSize
      in: query
      description: |
        The maximum number of results to return per page.
      example: 100
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 1000
    Cursor:
      name: cursor
      in: query
      description: |
        Parameter used in pagination requests. Maximum page size is set to 15.
        Set to the value of next for the next page of results.
        Set to the value of previous for the previous page of results.
        No other parameters can be set when this parameter is set.
      schema:
        type: string
        example: aHR0cHM6Ly9nLnBhZ2UvTmVrby1SYW1lbj9zaGFyZQ==
    RuleID:
      name: ruleID
      in: path
      required: true
      description: The rule ID.
      schema: { type: string, format: uuid }
...
  responses:
    NoContent:
      description: No content
...
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Rule:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RuleResponse' }
    Rules:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RulesCursorResponse' }
    Evaluation:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EvaluationResponse' }
...
  requestBodies:
    Rule:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RuleRequest' }
    RulePatch:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RulePatchRequest' }
...
  schemas:
    QueryBuilder:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          example: VALIDATION
        errorMessage:
          type: string
        details:
          type: string

    # --- V1 Ledger Clarity schemas -----------------------------------------
    Severity:
      type: string
      enum: [info, low, medium, high, critical]
    TemplateKind:
      type: string
      description: |
        The V1 GA template catalog. (`ledger_vs_pool_drift` was removed in 6a-2,
        subsumed by `source_parity`.)
      enum: [ledger_invariant, source_parity, account_threshold]
    Cadence:
      type: string
      description: |
        Reconciliation rhythm. Scopes each failing fingerprint into a period so a
        March break and an April break are distinct, independently-closable cases.
        `continuous` (default) is a single unbounded period (live monitoring).
      enum: [continuous, daily, weekly, monthly]
      default: continuous
    Schedule:
      type: object
      required: [kind]
      properties:
        kind: { type: string, enum: [on_demand, cron] }
        expr: { type: string, example: "*/15 * * * *" }
        tz: { type: string, example: "UTC" }
...
    Rule:
      type: object
      required: [id, name, templateKind, templateSpec, enabled, severity, cadence, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        templateKind: { $ref: '#/components/schemas/TemplateKind' }
        templateSpec:
          type: object
          additionalProperties: true
        compiledCEL: { type: string }
        enabled: { type: boolean }
        severity: { $ref: '#/components/schemas/Severity' }
        cadence: { $ref: '#/components/schemas/Cadence' }
        schedule: { $ref: '#/components/schemas/Schedule' }
        notifications:
          type: array
          items: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RuleRequest:
      type: object
      required: [name, templateKind, templateSpec]
      properties:
        name: { type: string }
        templateKind: { $ref: '#/components/schemas/TemplateKind' }
        templateSpec:
          type: object
          additionalProperties: true
        severity: { $ref: '#/components/schemas/Severity' }
        cadence: { $ref: '#/components/schemas/Cadence' }
        schedule: { $ref: '#/components/schemas/Schedule' }
        notifications:
          type: array
          items: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
        enabled: { type: boolean }
    RulePatchRequest:
      type: object
      description: Partial update — only fields supplied are applied.
      properties:
        name: { type: string }
        templateKind: { $ref: '#/components/schemas/TemplateKind' }
        templateSpec:
          type: object
          additionalProperties: true
        enabled: { type: boolean }
        severity: { $ref: '#/components/schemas/Severity' }
        schedule: { $ref: '#/components/schemas/Schedule' }
        notifications:
          type: array
          items: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
    RuleResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Rule' }
    RulesCursorResponse:
      type: object
      required: [cursor]
      properties:
        cursor:
          type: object
          required: [pageSize, hasMore, data]
          properties:
            pageSize: { type: integer, format: int64, minimum: 1, maximum: 1000 }
            hasMore: { type: boolean }
            previous: { type: string }
            next: { type: string }
            data:
              type: array
              items: { $ref: '#/components/schemas/Rule' }
    EvaluateRuleRequest:
      type: object
      properties:
        at: { type: string, format: date-time, description: "PIT for the evaluation. Defaults to now." }
    Evaluation:
      type: object
      required: [id, ruleID, startedAt, endedAt, result, createdAt]
      properties:
        id: { type: string, format: uuid }
        ruleID: { type: string, format: uuid }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        result: { type: string, enum: [PASS, FAIL, ERROR] }
        evidence:
          oneOf:
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: object
              additionalProperties: true
        error: { type: string }
        costUnits: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
    EvaluationResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Evaluation' }
...